Add mvi fragment delegate
Remove saving ViewState as parcelable
