+++ app/src/main/java/pl/valueadd/mvi/example/presentation/main/first/FirstViewState.kt
@@ -1,12 +1,10 @@
 package pl.valueadd.mvi.example.presentation.main.first
 
-import kotlinx.android.parcel.Parcelize
 import org.apache.commons.lang3.StringUtils.EMPTY
 import org.apache.commons.lang3.math.NumberUtils.INTEGER_ZERO
 import pl.valueadd.mvi.fragment.mvi.IBasePartialState
 import pl.valueadd.mvi.fragment.mvi.IBaseViewState
 
-@Parcelize
 data class FirstViewState(
     val count: Int = INTEGER_ZERO,
     val error: String = EMPTY,
