+++ app/src/main/java/pl/valueadd/mvi/example/presentation/main/root/RootViewState.kt
@@ -1,10 +1,8 @@
 package pl.valueadd.mvi.example.presentation.main.root
 
-import kotlinx.android.parcel.Parcelize
 import pl.valueadd.mvi.fragment.mvi.IBasePartialState
 import pl.valueadd.mvi.fragment.mvi.IBaseViewState
 
-@Parcelize
 class RootViewState :
     IBaseViewState {
