+++ mvi-valueadd/src/main/java/pl/valueadd/mvi/fragment/delegate/fragment/MviFragmentDelegate.kt
@@ -1,0 +1,10 @@
+package pl.valueadd.mvi.fragment.delegate.fragment
+
+interface MviFragmentDelegate {
+
+    fun onStart()
+
+    fun onStop()
+
+    fun onDestroy()
+}
