--- mvi-valueadd/src/main/java/pl/valueadd/mvi/fragment/mvi/IBaseMviFragment.kt
@@ -1,5 +1,0 @@
-package pl.valueadd.mvi.fragment.mvi
-
-interface IBaseMviFragment<P : BaseMviPresenter<*, *, *, *>> {
-    var presenter: P
-}
