+++ mvi-valueadd/src/main/java/pl/valueadd/mvi/fragment/mvi/IBaseViewState.kt
@@ -1,5 +1,3 @@
 package pl.valueadd.mvi.fragment.mvi
 
-import android.os.Parcelable
-
-interface IBaseViewState : Parcelable
+interface IBaseViewState
